refactor(cool-navi): extract NaviMenu component
