+++ src/components/NaviMenu/tests/index.test.js
@@ -1,0 +1,10 @@
+// import React from 'react'
+// import { shallow } from 'enzyme'
+
+// import NaviMenu from '../index'
+
+describe('TODO <NaviMenu />', () => {
+  it('Expect to have unit tests specified', () => {
+    expect(true).toEqual(true)
+  })
+})
